framework: add shader_compiler
The ShaderCompiler is a class that compiles a wgsl shader into a naga
Module, doing additional preprocessing.
For now the only supported preprocessing is the include command, which
has the following syntaxes:

//@include file_name (where file_name is a relative path)
//@include :definition_name

The first syntax searches file_name and pastes it into the source string,
while the second syntax search the given definition in the definitions
registered with the compiler.

A definition can be created using the ShaderCompiler::define() function