WIP implementation of undo/redo for chonky layers
At the moment there are some quirks to figure out, but the framework
is there.

Undo/redo on ChunkedLayers are performed through a ChunkDiff, which is
generated when edit() is called.

The ChunkedDiff can be applied to a layer through the
'apply_to_chunked_layer' function, which returns the inverted diff:
* if a chunk is created, in the inverted diff it's deleted
* if a chunk is modified, in the diff the version that revers to the
  current chunk texture is inserted
 * if a chunk is deleted, in the diff it's recreated (with the contents
   on deletion)

 To better work on undo/redo, since in the future it was planned anyway,
 stamping on ImageLayers has been removed